pyproject.toml
[tool.poetry]
name = "api-de-banco"
version = "0.1.0"
description = ""
authors = ["LeandroDeJesus-S <jstleandro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
databases = {extras = ["aiosqlite"], version = "^0.9.0"}
sqlalchemy = "^2.0.34"
pydantic = "^2.9.0"
pydantic-settings = "^2.4.0"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
mypy = "^1.11.2"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"

[tool.mypy]
exclude = [
    "migrations", 
    "tests"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -vvv"
testpaths = [
    "tests",
]
asyncio_mode = 'auto'
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
